+++ src/Compiler/NodeVisitor/Behavior/ScopeNodeVisitorInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Efabrica\PHPStanLatte\Compiler\NodeVisitor\Behavior;
+
+use PHPStan\Analyser\Scope;
+
+interface ScopeNodeVisitorInterface
+{
+    public function setScope(Scope $scope): void;
+}